Implement charge update funciton. Modify charge id type to long

+++ src/main/kotlin/com/example/point/domain/valueObjects/ChargedPoints.kt
@@ -3,12 +3,12 @@ package com.example.point.domain.valueObjects
 import kotlin.math.min
 
 class ChargedPoints(
-    val chargeId: Int,
+    val chargeId: Long,
     private val initPoint: Int,
 ) {
     private var usedPoints: Int = 0
 
-    class PointUsage(val chargeId: Int, val points: Int)
+    class PointUsage(val chargeId: Long, val points: Int)
 
     fun getLeftPoints(): Int {
         return initPoint - usedPoints
